requirements-flask.txt
flask>=2.3.0
spotipy>=2.22.1
python-dotenv>=1.0.0
gunicorn>=21.2.0